catalog_backend-master/src/pyproject.toml
[tool.poetry]
name = "hainu-backend"
version = "0.1.0"
description = ""
authors = ["jaklimoff <jaklimoff@gmail.com>"]
#readme = "README.md"
packages = [{ include = "hainu_backend" }]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.3"
djangorestframework = "^3.14.0"
markdown = "^3.4.4"
django-filter = "^23.2"
psycopg2-binary = "^2.9.6"
dj-database-url = "^2.0.0"
gunicorn = "^21.2.0"
whitenoise = "^6.5.0"
django-cors-headers = "^4.2.0"
django-tinymce = "^3.6.1"
six = "^1.16.0"
boto3 = "^1.28.35"
django-storages = "^1.13.2"
pillow = "^10.0.0"
django-environ = "^0.10.0"
httpx = "^0.24.1"
beautifulsoup4 = "^4.12.2"
drf-yasg = "^1.21.10"
requests = "^2.32.4"

[tool.poetry.dev-dependencies]
pytest = "^7.4.1"
factory-boy = "^3.3.0"
pytest-django = "^4.5.2"
pytest-env = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
